Clean up userspace loader and test app (#47)
* Move most debug code from the loader to the test app.
* Split the test app into separate files.
