+++ userspace/testapp/util.c
@@ -1,22 +1,22 @@
 /*
- * Copyright (C) 2019-2023 Philippe Aubertin.
+ * Copyright (C) 2023 Philippe Aubertin.
  * All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- * 
+ *
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 
+ *
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,49 +29,22 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <jinue/jinue.h>
-#include <jinue/util.h>
-#include <errno.h>
+
+#include <stdbool.h>
+#include <stddef.h>
 #include <stdlib.h>
-#include "tests/ipc.h"
-#include "debug.h"
+#include <string.h>
 #include "util.h"
 
-#define MAP_BUFFER_SIZE 16384
+bool bool_getenv(const char *name) {
+    const char *value = getenv(name);
 
-int main(int argc, char *argv[]) {
-    char call_buffer[MAP_BUFFER_SIZE];
-    int status;
-
-    /* Say hello. */
-    jinue_info("Jinue test app (%s) started.", argv[0]);
-
-    dump_cmdline_arguments(argc, argv);
-    dump_environ();
-    dump_auxvec();
-    dump_syscall_implementation();
-
-    /* get free memory blocks from microkernel */
-    status = jinue_get_user_memory((jinue_mem_map_t *)&call_buffer, sizeof(call_buffer), &errno);
-
-    if(status != 0) {
-        jinue_error("error: could not get physical memory map from microkernel");
-
-        return EXIT_FAILURE;
+    if(value == NULL) {
+        return false;
     }
 
-    dump_phys_memory_map((jinue_mem_map_t *)&call_buffer);
-
-    run_ipc_test();
-
-    if(bool_getenv("DEBUG_DO_REBOOT")) {
-        jinue_info("Rebooting.");
-        jinue_reboot();
-    }
-
-    while (1) {
-        jinue_yield_thread();
-    }
-    
-    return EXIT_SUCCESS;
+    return  strcmp(value, "enable") == 0 ||
+            strcmp(value, "true") == 0 ||
+            strcmp(value, "yes") == 0 ||
+            strcmp(value, "1") == 0;
 }
